requirements.txt
Django==1.7
jsonfield==1.0.0
djangorestframework
markdown
django-filter